Move Ixed into its own module.
